messageテーブルへのアクセスメソッドをModel_MessageからRepositoryへ切り出し
