+++ fuel/app/classes/controller/discover.php
@@ -15,7 +15,7 @@ class controller_discover extends Controller
 
     public function get_index()
     {
-        $messages = Model_Message::message_find_all();
+        $messages = Repository_Modelmessage::message_find_all();
 
         return View::forge('timeline', array(
             'aaa'      => Session::get('id'),
