+++ fuel/app/classes/controller/mentions.php
@@ -15,7 +15,7 @@ class Controller_mentions extends Controller
 
     public function get_index()
     {
-        $messages = Model_Message::message_find_by_mention_desc_by_createdAt(Session::get('id'));
+        $messages = Repository_Modelmessage::message_find_by_mention_desc_by_createdAt(Session::get('id'));
 
         return View::forge('timeline', array(
             'aaa'      => Session::get('id'),
